requirements.txt
numpy==1.15.4
XlsxWriter==1.1.2
Click==7.0
openpyxl==2.5.12
Flask==1.0.2
setuptools==40.6.3
xlwt==1.3.0
pytz==2018.7
Pillow==6.1.0
PyQt5==5.13.0
dlib==19.17.0
face_recognition_models==0.3.0
mysql_connector_repackaged==0.3.1
picamera==1.13
scikit_learn==0.21.2
